pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.8.1",
    "flask>=3.1.1",
    "huggingface-hub[mcp]>=0.33.0",
    "mcp[cli]>=1.9.4",
    "pywebview>=5.4",
]
